EL-86 Перейти на совместную генерацию пациента и родителя

+++ feature/login/src/main/java/ru/dekabrsky/login/data/mapper/LoginDataMapper.kt
@@ -25,7 +25,7 @@ class LoginDataMapper @Inject constructor() {
                 Base64.encode("$login:$password".toByteArray(), Base64.DEFAULT),
                 StandardCharsets.UTF_8
             ),
-            code = code.toInt()
+            code = code
         )
     }
